feat: 엔트리포인트 추가 및 rrf 병합 로직과 분리
